requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generative-ai
django
